+++ R/map.R
@@ -46,6 +46,17 @@
 #'   This makes it easier to understand which arguments belong to which
 #'   function and will tend to yield better error messages.
 #'
+#' @param .parallel `r lifecycle::badge("experimental")` Whether to map in
+#'   parallel. Use `TRUE` to parallelize using the \CRANpkg{mirai} package.
+#'   * Set up parallelization in your session beforehand using
+#'   [mirai::daemons()].
+#'   * Non-package functions are auto-crated for sharing with parallel
+#'   processes. You may [carrier::crate()] your function explicitly if you need
+#'   to supply additional objects along with your function.
+#'   * Use of `...` is not permitted in this context, [carrier::crate()] an
+#'   anonymous function instead.
+#'
+#'  See [parallelization] for more details.
 #' @param .progress Whether to show a progress bar. Use `TRUE` to turn on
 #'   a basic progress bar, use a string to give it a name, or see
 #'   [progress_bars] for more details.
@@ -125,50 +136,69 @@
 #'   map(\(df) lm(mpg ~ wt, data = df)) |>
 #'   map(summary) |>
 #'   map_dbl("r.squared")
-map <- function(.x, .f, ..., .progress = FALSE) {
-  map_("list", .x, .f, ..., .progress = .progress)
+#'
+#' @examplesIf interactive() && requireNamespace("mirai", quietly = TRUE) && requireNamespace("carrier", quietly = TRUE)
+#' # Run in interactive sessions only as spawns additional processes
+#'
+#' # To use parallelized map, set daemons (number of parallel processes) first:
+#' mirai::daemons(2)
+#'
+#' mtcars |> map_dbl(sum, .parallel = TRUE)
+#'
+#' 1:10 |>
+#'   map(function(x) stats::rnorm(10, mean = x), .parallel = TRUE) |>
+#'   map_dbl(mean, .parallel = TRUE)
+#'
+#' mirai::daemons(0)
+#'
+map <- function(.x, .f, ..., .parallel = FALSE, .progress = FALSE) {
+  map_("list", .x, .f, ..., .parallel = .parallel, .progress = .progress)
 }
 
 #' @rdname map
 #' @export
-map_lgl <- function(.x, .f, ..., .progress = FALSE) {
-  map_("logical", .x, .f, ..., .progress = .progress)
+map_lgl <- function(.x, .f, ..., .parallel = FALSE, .progress = FALSE) {
+  map_("logical", .x, .f, ..., .parallel = .parallel, .progress = .progress)
 }
 
 #' @rdname map
 #' @export
-map_int <- function(.x, .f, ..., .progress = FALSE) {
-  map_("integer", .x, .f, ..., .progress = .progress)
+map_int <- function(.x, .f, ..., .parallel = FALSE, .progress = FALSE) {
+  map_("integer", .x, .f, ..., .parallel = .parallel, .progress = .progress)
 }
 
 #' @rdname map
 #' @export
-map_dbl <- function(.x, .f, ..., .progress = FALSE) {
-  map_("double", .x, .f, ..., .progress = .progress)
+map_dbl <- function(.x, .f, ..., .parallel = FALSE, .progress = FALSE) {
+  map_("double", .x, .f, ..., .parallel = .parallel, .progress = .progress)
 }
 
 #' @rdname map
 #' @export
-map_chr <- function(.x, .f, ..., .progress = FALSE) {
+map_chr <- function(.x, .f, ..., .parallel = FALSE, .progress = FALSE) {
   local_deprecation_user_env()
-  map_("character", .x, .f, ..., .progress = .progress)
+  map_("character", .x, .f, ..., .parallel = .parallel, .progress = .progress)
 }
 
 map_ <- function(.type,
                  .x,
                  .f,
                  ...,
+                 .parallel = FALSE,
                  .progress = FALSE,
                  .purrr_user_env = caller_env(2),
                  .purrr_error_call = caller_env()) {
   .x <- vctrs_vec_compat(.x, .purrr_user_env)
   vec_assert(.x, arg = ".x", call = .purrr_error_call)
 
-  n <- vec_size(.x)
-
-  names <- vec_names(.x)
-
   .f <- as_mapper(.f, ...)
+
+  if (isTRUE(.parallel)) {
+    return(mmap_(.x, .f, .progress, .type, .purrr_error_call, ...))
+  }
+
+  n <- vec_size(.x)
+  names <- vec_names(.x)
 
   i <- 0L
   with_indexed_errors(
@@ -179,21 +209,62 @@ map_ <- function(.type,
   )
 }
 
+mmap_ <- function(.x, .f, .progress, .type, error_call, ...) {
+
+  if (is.null(the$packages_installed)) {
+    rlang::check_installed(c("mirai", "carrier"), reason = "for parallel map.")
+    the$packages_installed <- TRUE
+  }
+
+  if (is.null(mirai::nextget("n"))) {
+    cli::cli_abort(
+      "No daemons set - use e.g. {.run mirai::daemons(6)} to set 6 local daemons.",
+      call = error_call
+    )
+  }
+  if (...length()) {
+    cli::cli_abort(
+      "Don't use `...` with `.parallel = TRUE`.",
+      call = error_call
+    )
+  }
+
+  if (!isNamespace(topenv(environment(.f))) && !carrier::is_crate(.f)) {
+    .f <- carrier::crate(rlang::set_env(.f))
+    cli::cli_inform(c(
+      v = "Automatically crated `.f`: {format(lobstr::obj_size(.f))}"
+    ))
+  }
+
+  m <- mirai::mirai_map(.x, .f)
+
+  options <- c(".stop", if (isTRUE(.progress)) ".progress")
+  x <- with_parallel_indexed_errors(
+    mirai::collect_mirai(m, options = options),
+    interrupt_expr = mirai::stop_mirai(m),
+    error_call = error_call
+  )
+  if (.type != "list") {
+    x <- simplify_impl(x, ptype = vector(mode = .type), error_call = error_call)
+  }
+  x
+
+}
 
 #' @rdname map
 #' @param .ptype If `NULL`, the default, the output type is the common type
 #'   of the elements of the result. Otherwise, supply a "prototype" giving
 #'   the desired type of output.
 #' @export
-map_vec <- function(.x, .f, ..., .ptype = NULL, .progress = FALSE) {
-  out <- map(.x, .f, ..., .progress = .progress)
+map_vec <- function(.x, .f, ..., .ptype = NULL, .parallel = FALSE, .progress = FALSE) {
+  out <- map(.x, .f, ..., .parallel = .parallel, .progress = .progress)
   simplify_impl(out, ptype = .ptype)
 }
 
 #' @rdname map
 #' @export
-walk <- function(.x, .f, ..., .progress = FALSE) {
-  map(.x, .f, ..., .progress = .progress)
+walk <- function(.x, .f, ..., .parallel = FALSE, .progress = FALSE) {
+  map(.x, .f, ..., .parallel = .parallel, .progress = .progress)
   invisible(.x)
 }
 
@@ -221,6 +292,30 @@ with_indexed_errors <- function(expr, i, names = NULL, error_call = caller_env()
           class = "purrr_error_indexed"
         )
       }
+    }
+  )
+}
+
+with_parallel_indexed_errors <- function(expr, interrupt_expr = NULL, error_call = caller_env()) {
+  withCallingHandlers(
+    expr,
+    error = function(cnd) {
+      location <- cnd$location
+      iname <- cnd$name
+      cli::cli_abort(
+        c(
+          i = "In index: {location}.",
+          i = if (length(iname) && nzchar(iname)) "With name: {iname}."
+        ),
+        location = location,
+        name = iname,
+        parent = cnd$parent,
+        call = error_call,
+        class = "purrr_error_indexed"
+      )
+    },
+    interrupt = function(cnd) {
+      interrupt_expr
     }
   )
 }